refactor: :recycle: split rewinding function
