+++ src/quicksort/a_to_b.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   quicksort.c                                        :+:      :+:    :+:   */
+/*   a_to_b.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/16 11:55:35 by youkim            #+#    #+#             */
-/*   Updated: 2021/12/19 19:59:58 by youkim           ###   ########.fr       */
+/*   Updated: 2021/12/19 20:12:42 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,56 +46,6 @@ void	partition_initial(t_engine *e)
 	b_to_a( e, psize[SMOL]);
 }
 
-void	b_to_a(t_engine *e, int size)
-{
-	int	pivot[2];
-	int	psize[3];
-
-	if (size <= 2)
-	{
-		for (int i = 0; i < size; i++)
-			oper(e, STK_B, PUSH);
-		return ((void)smolsort(e, STK_A, size));
-		 ;
-	}
-	init_array(psize, 3);
-	set_pivot(e, STK_B, size, pivot);
-
-	while (--size >= 0)
-	{
-		if (is_big(e, STK_B, pivot))
-		{
-			oper(e, STK_B, PUSH);
-			psize[BIG]++;
-		}
-		else if (is_mid(e, STK_B, pivot))
-		{
-			oper(e, STK_B, PUSH);
-			oper(e, STK_A, ROT);
-			psize[MID]++;
-		}
-		else
-		{
-			oper(e, STK_B, ROT);
-			psize[SMOL]++;
-		}
-	}
-
-	a_to_b(e, psize[BIG]);
-
-	for (int i = 0; i < psize[MID]; i++) // rewind mid
-	{
-		oper(e, STK_A, RROT);
-	}
-	for (int i = 0; i < psize[SMOL]; i++) // rewind smol
-	{
-		oper(e, STK_B, RROT);
-	}
-
-	a_to_b(e, psize[MID]);
-	b_to_a(e, psize[SMOL]);
-}
-
 void	a_to_b(t_engine *e, int size)
 {
 	int	pivot[2];
@@ -124,11 +74,11 @@ void	a_to_b(t_engine *e, int size)
 			psize[SMOL]++;
 		}
 	}
-
-	for (int i = 0; i < psize[BIG]; i++)
-		oper(e, STK_A, RROT);
-	for (int i = 0; i < psize[MID]; i++)
-		oper(e, STK_B, RROT);
+	rewind_sector(e, psize, (t_sortflag [2]){BIG, MID});
+	// for (int i = 0; i < psize[BIG]; i++)
+	// 	oper(e, STK_A, RROT);
+	// for (int i = 0; i < psize[MID]; i++)
+	// 	oper(e, STK_B, RROT);
 
 	a_to_b(e, psize[BIG]); // [3]
 	b_to_a(e, psize[MID]); // [2]
